pyproject.toml
[tool.poetry]
name = "bbtrack-yolo"
version = "0.1.0"
description = ""
authors = ["Jonathan Yang <hi@jyang.cc>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "bbtrack_yolo", from = "." }]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.2.1"
torchvision = "^0.17.1"
ultralytics = "^8.1.23"
scikit-learn = "^1.4.1.post1"
loguru = "^0.7.2"
tensorboard = "^2.16.2"
comet-ml = "^3.38.1"
pyarrow = "^15.0.0"
labelbox = { extras = ["data"], version = "^3.63.0" }
lapx = "^0.5.5"
pandera = "^0.18.0"
pydantic = "^2.6.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
mypy = "^1.8.0"
isort = "^5.13.2"
black = "^24.2.0"
pandas-stubs = "^2.2.0.240218"
types-tqdm = "^4.66.0.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
